S_KaladaiYT/Level2_QuizPrep/Session_03_FunctionCalling_Tools/tools_functionCalling/pyproject.toml
[project]
name = "tools-functioncalling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents==0.2.11",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
